requirements.txt
accelerate==0.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
dill==0.3.9
faiss-cpu==1.9.0.post1
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
inquirerpy==0.3.4
numpy==2.1.3
packaging==24.2
pandarallel==1.6.5
pandas==2.2.3
pfzy==0.3.4
pillow==11.0.0
prompt_toolkit==3.0.48
psutil==6.1.0
PyMuPDF==1.24.14
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
six==1.16.0
tokenizers==0.20.3
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
